Fix dependencies: Python 3.11.9 breaks Dask
Python 3.11.9 started to employ a different behaviour when raising
`TypeError`s when reflecting docstrings, which leads to that error when
importing `dask.dataframe`.

  TypeError: descriptor '__call__' for 'type' objects doesn't apply to
  a 'property' object

The solution is either to limit execution to Python <3.11.9, or to
upgrade the `dask` and `distributed` packages to their most recent
versions.

Because Fugue can't control on which Python versions it is used
downstream, this patch uses the option to upgrade Dask.

Because GitHub recently rolled out Python 3.11.9 on their runners, and
other downstream users may also upgrade soon, it is advisable to release
this update on behalf of a bugfix release, to avoid things going south
for users of Fugue.